+++ spec/models/activity_spec.rb
@@ -420,7 +420,7 @@ RSpec.describe Activity, type: :model do
 
     context "when gcrf_strategic_area has too many values" do
       let(:source_fund_code) { Fund.by_short_name("NF").id }
-      let(:strategic_areas) { %w[RF Clm IP] }
+      let(:strategic_areas) { %w[RF CIm IP] }
       subject { build(:programme_activity, source_fund_code: source_fund_code, gcrf_strategic_area: strategic_areas) }
 
       context "with a GCRF funded activity" do